requirements.txt
# Flask and web dependencies
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7

# Celery and Redis
celery==5.3.4
redis==5.0.1
flower==2.0.1

# Medical image processing
nibabel==5.2.0
pydicom==2.4.3
numpy==1.24.4
Pillow==10.0.1

# Database and storage
supabase==2.0.2

# Utilities
python-dotenv==1.0.0
gunicorn==20.1.0
tensorflow==2.14.0
# Note: The following are built-in Python modules and don't need to be installed:
# uuid, datetime, hashlib, shutil, tempfile, logging, threading, time, 
# functools, concurrent.futures, queue, json, os